Add a way to purge temp buffer data (#89)
If serialization is being canceled, there should be a way to purge the
temp buffer (clear memory / delete temp file) that is being used.
bump to 2.4.3

+++ src/node/warcserializer.ts
@@ -75,7 +75,14 @@ export class TempFileBuffer extends warcserializer.SerializerInMemBuffer {
       yield buff;
     }
 
-    await unlink(this.filename);
+    await this.purge();
+  }
+
+  override async purge() {
+    if (this.filename) {
+      await unlink(this.filename);
+      this.filename = "";
+    }
   }
 }
